refact: check if deletion warning file already exists before create it

+++ src/infra/services/servicesCmdRepo.go
@@ -250,8 +250,12 @@ func (repo *ServicesCmdRepo) createDefaultDirectories(
 		}
 
 		deletionWarningFilePath := defaultDirPath + "/DONOTDELETE"
+		if infraHelper.FileExists(deletionWarningFilePath) {
+			continue
+		}
+
 		_, err = os.Create(deletionWarningFilePath)
-		if err != nil && !os.IsExist(err) {
+		if err != nil {
 			return errors.New("CreateDeletionWarningFileError: " + err.Error())
 		}
 	}
